routing: Use 64-bit msatoshi for messages to and from routing.
Internally both payment and routing use 64-bit, but the interface
between them used 32-bit.
Since both components already support 64-bit we should use that.

+++ gossipd/routing.h
@@ -176,7 +176,7 @@ get_channel(const struct routing_state *rstate,
 struct route_hop {
 	struct short_channel_id channel_id;
 	struct pubkey nodeid;
-	u32 amount;
+	u64 amount;
 	u32 delay;
 };
 
@@ -238,7 +238,7 @@ struct node *get_node(struct routing_state *rstate, const struct pubkey *id);
 struct route_hop *get_route(const tal_t *ctx, struct routing_state *rstate,
 			    const struct pubkey *source,
 			    const struct pubkey *destination,
-			    const u32 msatoshi, double riskfactor,
+			    const u64 msatoshi, double riskfactor,
 			    u32 final_cltv,
 			    double fuzz,
 			    const struct siphash_seed *base_seed);
